create EventAttendee POJO with embedded field to handle composite key
Java won't like it if I have two fields annotated with @Id, so this is
JTA's work around
