+++ src/com/examples/ezoo/model/EventAttendee.java
@@ -1,51 +1,34 @@
 package com.examples.ezoo.model;
 
-import java.io.Serializable;
+import javax.persistence.EmbeddedId;
+import javax.persistence.Entity;
+import javax.persistence.Table;
 
-import javax.persistence.Embeddable;
-import javax.validation.constraints.NotNull;
+@Entity
+@Table(name="EVENT_ATTENDEES")
+public class EventAttendee {
 
-@Embeddable
-public class EventAttendee implements Serializable{
-
-	private static final long serialVersionUID = 1L;
-
-	@NotNull
-	private String username;
+	@EmbeddedId
+	private EmbeddedEventAttendee eventAttendee = new EmbeddedEventAttendee();
 	
-	@NotNull(message = "{eventid.validate}")
-	private Integer eventID;
+	public EventAttendee() {}
 	
-	public String getUsername() {
-		return username;
+	public EventAttendee(EmbeddedEventAttendee eventAttendee) {
+		this.eventAttendee = eventAttendee;
 	}
 
-	public void setUsername(String username) {
-		this.username = username;
+	public EmbeddedEventAttendee getEventAttendee() {
+		return eventAttendee;
 	}
 
-	public Integer getEventID() {
-		return eventID;
+	public void setEventAttendee(EmbeddedEventAttendee eventAttendee) {
+		this.eventAttendee = eventAttendee;
 	}
 
-	public void setEventID(Integer eventID) {
-		this.eventID = eventID;
-	}
-
-	@Override public boolean equals(Object o) {
-		if (this == o) return true;
-		if (o == null || this.getClass() != o.getClass()) return false;
-		
-		EventAttendee that = (EventAttendee) o;
-		
-		if (!username.equals(that.username)) return false;
-		return eventID.equals(that.eventID);
-	}
-	
 	@Override
-	public int hashCode() {
-		int result = username.hashCode();
-		result = 31 * result + eventID.hashCode();
-		return result;
+	public String toString() {
+		return "EventAttendee [eventAttendee=" + eventAttendee + "]";
 	}
+	
+	
 }
